pyproject.toml
[project]
name = "chip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "groq>=0.31.0",
    "httpx>=0.28.1",
    "hypercorn==0.14.4",
    "openai>=1.99.9",
    "orjson>=3.11.2",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
    "tiktoken>=0.11.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.16",
    "pytest>=8.4.1",
    "respx>=0.22.0",
    "ruff>=0.12.8",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
]
